UserProfile type
