+++ api/models/User.ts
@@ -1,6 +1,12 @@
 import { z } from "zod"
 import { UserSettings, UserSettingsSchema } from "../../domain-models/Settings"
-import { FriendRequestSentStatusSchema, FriendsStatusSchema, UserHandleSchema, UserIDSchema, UserToProfileRelationStatusSchema } from "../../domain-models/User"
+import {
+  FriendRequestSentStatusSchema,
+  FriendsStatusSchema,
+  UserHandleSchema,
+  UserIDSchema,
+  UserToProfileRelationStatusSchema
+} from "../../domain-models/User"
 import { tifAPIErrorSchema } from "./Error"
 
 export const userTiFAPIErrorSchema = <const T extends string>(literal: T) => {
@@ -12,8 +18,7 @@ export const userTiFAPIErrorSchema = <const T extends string>(literal: T) => {
 export const UserNotFoundResponseSchema =
   userTiFAPIErrorSchema("user-not-found")
 
-export const BlockedUserResponseSchema =
-  userTiFAPIErrorSchema("blocked")
+export const BlockedUserResponseSchema = userTiFAPIErrorSchema("blocked")
 
 export const UpdateCurrentUserProfileRequestSchema = z.object({
   name: z.string().optional(),
@@ -37,7 +42,12 @@ export type UserSettingsResponse = UserSettings
 
 export const UserSettingsResponseSchema = UserSettingsSchema
 
-export const UserFriendRequestResponseSchema = z.object({relationStatus: z.enum([FriendRequestSentStatusSchema.value, FriendsStatusSchema.value])})
+export const UserFriendRequestResponseSchema = z.object({
+  relationStatus: z.enum([
+    FriendRequestSentStatusSchema.value,
+    FriendsStatusSchema.value
+  ])
+})
 
 const DevicePlatformSchema = z.enum(["apple", "android"])
 
@@ -48,20 +58,8 @@ export const SelfProfileSchema = z.object({
   handle: UserHandleSchema,
   createdDateTime: z.coerce.date(),
   profileImageURL: z.string().url().optional(),
-  updatedDateTime: z.coerce.date(),
+  updatedDateTime: z.coerce.date()
 })
-
-export const UserProfileSchema = z.object({
-  id: UserIDSchema,
-  name: z.string(),
-  bio: z.string().optional(),
-  handle: UserHandleSchema,
-  createdDateTime: z.coerce.date(),
-  profileImageURL: z.string().url().optional(),
-  updatedDateTime: z.coerce.date(),
-  relationStatus: UserToProfileRelationStatusSchema
-})
-
 export const RegisterPushTokenRequestSchema = z.object({
   pushToken: z.string().min(1), //generic nonempty string schema?
   platformName: DevicePlatformSchema
